@@ -5,7 +5,7 @@ declare(strict_types=1);
 /*
  * This file is part of Flight Routing.
  *
- * PHP version 7.1 and above required
+ * PHP version 7.4 and above required
  *
  * @author    Divine Niiquaye Ibok <divineibok@gmail.com>
  * @copyright 2019 Biurad Group (https://biurad.com/)
@@ -25,7 +25,7 @@ use Flight\Routing\Interfaces\ExceptionInterface;
 class MethodNotAllowedException extends \RuntimeException implements ExceptionInterface
 {
     /** @var string[] */
-    private $methods;
+    private array $methods;
 
     /**
      * @param string[] $methods
