@@ -5,7 +5,7 @@ declare(strict_types=1);
 /*
  * This file is part of Flight Routing.
  *
- * PHP version 7.1 and above required
+ * PHP version 7.4 and above required
  *
  * @author    Divine Niiquaye Ibok <divineibok@gmail.com>
  * @copyright 2019 Biurad Group (https://biurad.com/)
@@ -27,8 +27,7 @@ final class ResourceHandler
     /** @var string|object */
     private $classResource;
 
-    /** @var string */
-    private $actionResource;
+    private string $actionResource;
 
     /**
      * @param class-string|object $class  of class string or class object
