@@ -5,7 +5,7 @@ declare(strict_types=1);
 /*
  * This file is part of Flight Routing.
  *
- * PHP version 7.1 and above required
+ * PHP version 7.4 and above required
  *
  * @author    Divine Niiquaye Ibok <divineibok@gmail.com>
  * @copyright 2019 Biurad Group (https://biurad.com/)
@@ -36,8 +36,7 @@ use PHPUnit\Framework\TestCase;
  */
 class ListenerTest extends TestCase
 {
-    /** @var AnnotationLoader */
-    protected $loader;
+    protected AnnotationLoader $loader;
 
     protected function setUp(): void
     {
@@ -529,7 +528,7 @@ class ListenerTest extends TestCase
                 'defaults' => ['foo' => 'bar'],
                 'patterns' => ['locale' => 'en|fr'],
                 'arguments' => [],
-            ]
+            ],
         ], Fixtures\Helper::routesToArray($routes));
     }
 
