@@ -5,7 +5,7 @@ declare(strict_types=1);
 /*
  * This file is part of Flight Routing.
  *
- * PHP version 7.1 and above required
+ * PHP version 7.4 and above required
  *
  * @author    Divine Niiquaye Ibok <divineibok@gmail.com>
  * @copyright 2019 Biurad Group (https://biurad.com/)
@@ -23,12 +23,12 @@ use Flight\Routing\Annotation\Route;
 class GlobalDefaultsClass
 {
     #[Route(path: '/specific-name', name: 'specific_name', methods: ['GET'])]
-    public function withName()
+    public function withName(): void
     {
     }
 
     #[Route('/specific-none', methods: ['GET', 'HEAD'])]
-    public function noName()
+    public function noName(): void
     {
     }
 }
