@@ -5,7 +5,7 @@ declare(strict_types=1);
 /*
  * This file is part of Flight Routing.
  *
- * PHP version 7.1 and above required
+ * PHP version 7.4 and above required
  *
  * @author    Divine Niiquaye Ibok <divineibok@gmail.com>
  * @copyright 2019 Biurad Group (https://biurad.com/)
@@ -22,7 +22,7 @@ use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
 /**
- * BlankController
+ * BlankController.
  */
 class BlankController
 {
@@ -36,9 +36,6 @@ class BlankController
      */
     private $attributes = [];
 
-    /**
-     * @return bool
-     */
     public function isRunned(): bool
     {
         return $this->isRunned;
@@ -62,12 +59,9 @@ class BlankController
         return $this->attributes[$key] ?? null;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function handle(ServerRequestInterface $request): ResponseInterface
     {
-        $this->isRunned   = true;
+        $this->isRunned = true;
         $this->attributes = $request->getAttributes();
 
         return (new Psr17Factory())->createResponse();
