@@ -5,7 +5,7 @@ declare(strict_types=1);
 /*
  * This file is part of Flight Routing.
  *
- * PHP version 7.1 and above required
+ * PHP version 7.4 and above required
  *
  * @author    Divine Niiquaye Ibok <divineibok@gmail.com>
  * @copyright 2019 Biurad Group (https://biurad.com/)
@@ -21,15 +21,11 @@ use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
 /**
- * InvokeController
+ * InvokeController.
  */
 class InvokeController extends BlankController
 {
     /**
-     * @param ServerRequestInterface $request
-     *
-     * @return ResponseInterface
-     *
      * @link https://www.php.net/manual/en/language.oop5.magic.php#object.invoke
      */
     public function __invoke(ServerRequestInterface $request): ResponseInterface
